// Frontend/src/app/shared/model/patient.ts
export class Patient {
    PatientId: number = 0;
    PatientName: string = '';
    Dob: Date = new Date();
    Gender: string = '';
    BloodGroup: string = '';
    PatientPhone: string = '';
    PatientAddress: string = '';
    RegistrationDate: Date = new Date();
    IsActive: boolean = false;
}